Updated movie categories logic

+++ src/movies/movies.service.ts
@@ -5,11 +5,12 @@ import { PrismaService } from 'src/common/services/prisma.service'
 import { MovieCategoriesService } from 'src/movie-categories/movie-categories.service'
 import { PaginationArgs } from 'src/common/dto/pagination-args.dto'
 import { MoviePreview } from './entities/movie-preview.entity'
-import { MovieWithCategories } from './entities/movie-with-categories'
 import { CacheService } from '../common/services/cache.service'
 import { CacheKeys } from 'src/common/helpers/cache-keys.helper'
-import { handle } from 'src/common/utilities/handle.utility'
 import { catchTo } from 'src/common/utilities/catch-to.utility'
+import { tryTo } from 'src/common/utilities/try-to.utility'
+import { MovieWithCategories } from './entities/movie-with-categories'
+import { MovieWithRowCategories } from './entities/movie-with-row-categories.entity'
 
 @Injectable()
 export class MoviesService {
@@ -20,17 +21,13 @@ export class MoviesService {
   ) {}
 
   async create(data: CreateMovieDto) {
-    try {
-      const { categories, ...rest } = data
-      const newMovie = await this.prismaService.movie.create({ data: rest })
+    const { categories, ...rest } = data
+    const [newMovie] = await tryTo(this.prismaService.movie.create({ data: rest }))
 
-      if (categories) {
-        return await this.movieCategoriesService.addToMovie(newMovie, categories)
-      } else {
-        return newMovie
-      }
-    } catch (error) {
-      handle(error)
+    if (categories && newMovie) {
+      return await catchTo(this.movieCategoriesService.addToMovie(newMovie, categories))
+    } else {
+      return newMovie
     }
   }
 
@@ -42,13 +39,7 @@ export class MoviesService {
         this.prismaService.paginate<MoviePreview>({
           model: 'movie',
           paginationArgs,
-          options: {
-            select: {
-              title: true,
-              id: true,
-              thumbnail: true
-            }
-          }
+          options: { select: { title: true, id: true, thumbnail: true } }
         })
     })
   }
@@ -58,27 +49,19 @@ export class MoviesService {
       key: CacheKeys.PAGINATED_MOVIES(paginationArgs),
       ttl: '1h',
       fn: async () => {
-        const { data, nextCursor } = await this.prismaService.paginate<MovieWithCategories>({
-          model: 'movie',
-          paginationArgs,
-          options: {
-            orderBy: { releaseDate: 'asc' },
-            include: {
-              categories: {
-                include: {
-                  category: true
-                }
-              }
+        const { data: movies, nextCursor } =
+          await this.prismaService.paginate<MovieWithRowCategories>({
+            model: 'movie',
+            paginationArgs,
+            options: {
+              orderBy: { releaseDate: 'asc' },
+              include: { categories: { include: { category: true } } }
             }
-          }
-        })
+          })
 
         return {
-          data: data.map((movie) => ({
-            ...movie,
-            categories: movie.categories.map((c) => c.category.name)
-          })),
-          nextCursor
+          nextCursor,
+          data: movies.map((movie) => this._getWithCategories(movie))
         }
       }
     })
@@ -91,20 +74,11 @@ export class MoviesService {
       fn: async () => {
         const movie = await this.prismaService.movie.findUnique({
           where: { id },
-          include: {
-            categories: {
-              include: {
-                category: true
-              }
-            }
-          }
+          include: { categories: { include: { category: true } } }
         })
 
         if (movie) {
-          return {
-            ...movie,
-            categories: movie.categories.map((mc) => mc.category.name)
-          }
+          return this._getWithCategories(movie)
         } else {
           return null
         }
@@ -113,23 +87,28 @@ export class MoviesService {
   }
 
   async update(id: number, data: UpdateMovieDto) {
-    try {
-      const { categories, ...rest } = data
-      const updatedMovie = await this.prismaService.movie.update({ where: { id }, data: rest })
+    const { categories, ...rest } = data
+    const [updatedMovie] = await tryTo(
+      this.prismaService.movie.update({ where: { id }, data: rest })
+    )
 
-      if (categories) {
-        await this.movieCategoriesService.delete(id)
+    if (categories && updatedMovie) {
+      await this.movieCategoriesService.delete(id)
 
-        return await this.movieCategoriesService.addToMovie(updatedMovie, categories)
-      } else {
-        return updatedMovie
-      }
-    } catch (error) {
-      handle(error)
+      return await catchTo(this.movieCategoriesService.addToMovie(updatedMovie, categories))
+    } else {
+      return updatedMovie
     }
   }
 
   async delete(id: number) {
     return await catchTo(this.prismaService.movie.delete({ where: { id } }))
+  }
+
+  private _getWithCategories(movie: MovieWithRowCategories) {
+    const { categories, ...rest } = movie
+    const finalCategories = categories.map((category) => category.categoryName)
+
+    return { ...rest, categories: finalCategories } as MovieWithCategories
   }
 }
